Create Get-UserWhoBuiltEC2Instance.ps1
This PowerShell script uses the AWS PowerShell Module as well as the Splunk Module. I might change this later to use AWS Config so there's no reliance on Splunk. The other part this checks is it considers you are using some form of federated identity for your users who are launching your EC2 Instances (hence the split on the ARN which specifies Secure Token Service.